removed stl includes from commoninclude

+++ WickedEngine/wiRectPacker.cpp
@@ -252,7 +252,7 @@ namespace wiRectPacker
 			bins.push_back(bin());
 			b = &bins[bins.size() - 1];
 
-			b->size = _rect2D(&((*p[0])[0]), p[0]->size(), max_s, b->rects, *p[1]);
+			b->size = _rect2D(&((*p[0])[0]), static_cast<int>(p[0]->size()), max_s, b->rects, *p[1]);
 			b->rects.shrink_to_fit();
 			p[0]->clear();
